refactor: 누락된 export 반영
