+++ src/pages/ArticlesPage/RenderFollowingArticles.tsx
@@ -1,9 +1,7 @@
 import { useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useInfiniteQuery } from '@tanstack/react-query';
-import Loader from '@components/Loader';
-import SearchSkeleton from '@components/SearchSkeleton';
-import SubButton from '@components/SubButton';
+import { Loader, SearchSkeleton, SubButton } from '@/components';
 import { FOLLOWING_ARTICLE_FETCH_LIMIT } from '@constants/Article';
 import { TAB_CONSTANTS } from '@constants/Tab';
 import useAuthQuery from '@hooks/useAuthQuery';
