+++ src/pages/ArticlesPage/RenderNewestArticles.tsx
@@ -1,7 +1,6 @@
 import { useCallback } from 'react';
 import { useInfiniteQuery } from '@tanstack/react-query';
-import Loader from '@components/Loader';
-import SearchSkeleton from '@components/SearchSkeleton';
+import { Loader, SearchSkeleton } from '@/components';
 import { ARTICLE_FETCH_LIMIT } from '@constants/Article';
 import { TAB_CONSTANTS } from '@constants/Tab';
 import { fetchArticles } from './fetchArticles';
