+++ src/pages/NewArticlePage.tsx
@@ -5,9 +5,8 @@ import { User } from '@type/User';
 import { AiOutlineCloseCircle } from 'react-icons/ai';
 import { BiImageAdd } from 'react-icons/bi';
 import { Avatar, CloseButton, HeaderText, MainButton } from '@/components';
-import { useCreateArticle, useToastContext } from '@/hooks';
+import { useAuthQuery, useCreateArticle, useToastContext } from '@/hooks';
 import { DROPDOWN_OPTIONS, ETC, LENGTH_LIMIT, MESSAGE } from '@constants/NewArticle';
-import useAuthQuery from '@hooks/useAuthQuery';
 
 export type FormValueType = {
   title: string;
