+++ src/type/index.ts
@@ -1,0 +1,11 @@
+export * from './Channel';
+export * from './Comment';
+export * from './Conversation';
+export * from './Follow';
+export * from './Like';
+export * from './Message';
+export * from './Notification';
+export * from './Post';
+export * from './User';
+export * from './search';
+export * from './signUp';
